Implemented new custom post type - Press Releases
Created and implemented the custom WordPress Post Type for Press
Release at the theme level.
This implementation is very crude because this theme is on borrowed
time. The CSS for the Archive and Single Press Release is embed directly
in the template instead of using a deciated CSS File. (Forgive me)